Created separate modules for every class
